refactor : 마이페이지에서 사용되는 api 함수 하나의 파일로 이동 & index.ts 정의
