+++ src/service/api/my/getMyMentoringTeam.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import MentoringTeam from "@/types/team/mentoring/mentoringTeam";
-import { server } from "../../instance/server/server";
+import { server } from "../instance/server";
 
 const getMyMentoringTeam = async (): Promise<MentoringTeam[]> => {
   const response = await server.get("/users/mentoring/teams");
