+++ src/service/api/my/getMyProjectTeam.ts
@@ -1,6 +1,6 @@
 "use server";
 import ProjectTeam from "@/types/team/project/projectTeam";
-import { server } from "../../instance/server/server";
+import { server } from "../instance/server";
 
 const getMyProjectTeam = async (): Promise<ProjectTeam[]> => {
   const response = await server.get("/user/project");
