+++ src/service/api/my/getMyReview.ts
@@ -1,0 +1,9 @@
+import { client } from "../instance/client";
+
+import UserReview from "@/types/user/userReview";
+
+export const getMyReviews = async (): Promise<UserReview[]> => {
+  const response = await client.get("/mock/users/reviews");
+
+  return response.data.data;
+};
