+++ src/service/api/my/index.ts
@@ -1,0 +1,10 @@
+import { getMyInfo } from "./getMyInfo";
+import getMyMentoringTeam from "./getMyMentoringTeam";
+import getMyProjectTeam from "./getMyProjectTeam";
+import { getMyReviews } from "./getMyReview";
+
+// 회원정보
+export { getMyReviews, getMyInfo };
+
+// 팀 리스트
+export { getMyMentoringTeam, getMyProjectTeam };
